add info on chance of defeat

+++ src/ascii.cpp
@@ -303,7 +303,7 @@ void ascii_putf_internal2(int x, int y, color fg, color bg, const std::string &t
   auto  text_iter = text.begin();
 
   if (compiler_unused) {
-    printf("ascii_putf_internal2 [%s]/%ld scissors x %d y %d scissors %d %d %d %d %d\n", text.c_str(), text.size(), x, y, scissors_tl.x,
+    printf("ascii_putf_internal2 [%s]/%d scissors x %d y %d scissors %d %d %d %d %d\n", text.c_str(), (int) text.size(), x, y, scissors_tl.x,
            scissors_tl.y, scissors_br.x, scissors_br.y, scissors_enabled);
   }
